Moved record functions to own helper
